Moved isoform class to new file. Check if mutation can be mapped (to be tested).
